requirements.txt
astroid==2.3.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.1
dataclasses==0.8
dotmap==1.3.23
firebase-admin==4.5.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-secret-manager==2.0.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.2
libcst==0.3.18
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.23
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.4.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
typed-ast==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==16.7.5
Werkzeug==1.0.1
wrapt==1.11.2